Move Lowercase and NonEmptyString Attributes to validation folder
