+++ FU.API/FU.API/Validation/LowercaseAttribute.cs
@@ -1,4 +1,4 @@
-namespace FU.API.Helpers;
+namespace FU.API.Validation;
 
 using System.ComponentModel.DataAnnotations;
 
@@ -10,20 +10,6 @@ public class LowercaseAttribute : ValidationAttribute
         if (value is string && ((string)value).ToLower() != (string)value)
         {
             return new ValidationResult("String must be lowercase");
-        }
-
-        return ValidationResult.Success;
-    }
-}
-
-public class NonEmptyStringAttribute : ValidationAttribute
-{
-    protected override ValidationResult? IsValid(
-        object? value, ValidationContext validationContext)
-    {
-        if (value is string && (string)value == string.Empty)
-        {
-            return new ValidationResult("String can't be empty");
         }
 
         return ValidationResult.Success;
